pyproject.toml
[project]
name = "fastapi-template-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Tomás Vallotton",email = "tomasvallotton@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "typer (>=0.15.1,<0.16.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aiosmtplib (>=3.0.2,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "aioboto3 (>=13.3.0,<14.0.0)",
    "faker (>=33.3.1,<34.0.0)"
]

[tool.poetry]
packages = [{include = "src"}]


[tool.pyright]
reportIncompatibleVariableOverride = 'none'

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.2"
watchfiles = "^1.0.4"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
websockets = "^14.1"
beautifulsoup4 = "^4.12.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
